redesign codeblock
