+++ src/components/EnteSpinner.tsx
@@ -1,20 +1,6 @@
 import React from 'react';
-import { Spinner } from 'react-bootstrap';
+import CircularProgress from '@mui/material/CircularProgress';
 
 export default function EnteSpinner(props) {
-    const { style, ...others } = props ?? {};
-    return (
-        <Spinner
-            animation="border"
-            style={{
-                width: '36px',
-                height: '36px',
-                borderWidth: '0.20em',
-                color: '#51cd7c',
-                ...(style && style),
-            }}
-            {...others}
-            role="status"
-        />
-    );
+    return <CircularProgress color="accent" size={32} {...props} />;
 }
