+++ src/types/theme/styled.d.ts
@@ -1,0 +1,7 @@
+// styled.d.ts
+import 'styled-components';
+import { Theme } from '@mui/material';
+
+declare module 'styled-components' {
+    export interface DefaultTheme extends Theme {}
+}
